+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * core/json: KClass.createFrom will report missing parameters in a nicer way
 * server: introduced FormDataRenderer
 * core: conversion of enum values from string is now case-insensitive
+* json: can convert number to a custom type using opts.values
 
 # 1.6.1
 * jdbc: and() introduced in addition to or() for more convenient composition of where expressions